pyproject.toml
[tool.poetry]
name = "spam-bot-discord"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
praw = "^7.1.0"
flask = "^1.1.2"
dblpy = "^0.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
